refactor(tools): extract registerMetadataTools into tools/metadata.js
Phase C item 3: moves 7 sidecar-writing tools (create_character_sheet,
create_place_sheet, upsert_thread_link, update_scene_metadata,
update_character_sheet, update_place_sheet, flag_scene) out of index.js.
createCanonicalWorldEntity is passed through toolContext to avoid circular
imports. Removes writeMeta, normalizeSceneMetaForPath, and validateUniverseId
from index.js imports as they are no longer used directly there.
